ml_trading_bot/requirements.txt
ccxt
pandas
numpy
scikit-learn
joblib
python-dotenv
